+++ processing_routes/index.js
@@ -1,0 +1,43 @@
+const { onStreamingStart } = require('./onStreamingStart')
+const { onBlock } = require('./onBlock')
+const { send } = require('./send')
+const { power_up, power_down } = require('./power')
+const { delegate_vesting_shares } = require('./delegate_vesting_shares')
+const { vote, vote_content } = require('./vote')
+const { cert } = require('./cert')
+const { sig } = require('./sig')
+const { cjv } = require('./cjv')
+const { nomention } = require('./nomention')
+const { q4d } = require('./q4d')
+const { node_add, node_delete } = require('./nodes')
+const { dex_buy, dex_clear, dex_hbd_sell, dex_hive_sell, escrow_approve, escrow_dispute, escrow_release, escrow_transfer, transfer } = require('./dex')
+const { comment, comment_options } = require('./comment')
+
+exports = {
+    cert,
+    cjv,
+    comment,
+    comment_options,
+    delegate_vesting_shares,
+    dex_buy,
+    dex_clear,
+    dex_hbd_sell,
+    dex_hive_sell,
+    escrow_approve,
+    escrow_dispute,
+    escrow_release,
+    escrow_transfer,
+    node_add,
+    node_delete,
+    nomention,
+    onStreamingStart,
+    onBlock,
+    power_down,
+    power_up,
+    q4d,
+    send,
+    sig,
+    transfer,
+    vote,
+    vote_content
+}
