+++ processing_routes/q4d.js
@@ -1,0 +1,16 @@
+const { store, config } = require('./../index')
+
+exports.q4d = function(json, from, active, pc) {
+    if (from = config.leader && json.text && json.title) {
+        store.batch([{
+            type: 'put',
+            path: ['postQueue', json.title],
+            data: {
+                text: json.text,
+                title: json.title
+            }
+        }], pc)
+    } else {
+        pc[0](pc[2])
+    }
+}
