+++ processing_routes/send.js
@@ -1,0 +1,24 @@
+const { store, config } = require("./../index");
+const { getPathNum } = require("./../getPathNum");
+
+exports.send = function(json, from, active, pc) {
+    let fbalp = getPathNum(['balances', from]), //from balance promise
+        tbp = getPathNum(['balances', json.to]) //to balance promise
+    Promise.all([fbalp, tbp])
+        .then(bals => {
+            let fbal = bals[0], //from balance
+                tbal = bals[1], //to balance
+                ops = []
+            send = parseInt(json.amount)
+            if (json.to && typeof json.to == 'string' && send >= 0 && fbal >= send && active) { //balance checks
+                ops.push({ type: 'put', path: ['balances', from], data: parseInt(fbal - send) })
+                ops.push({ type: 'put', path: ['balances', json.to], data: parseInt(tbal + send) })
+                ops.push({ type: 'put', path: ['feed', `${json.block_num}:${json.transaction_id}`], data: `@${from}| Sent @${json.to} ${parseFloat(parseInt(json.amount)/1000).toFixed(3)} ${config.TOKEN}` })
+            } else {
+                ops.push({ type: 'put', path: ['feed', `${json.block_num}:${json.transaction_id}`], data: `@${from}| Invalid send operation` })
+            }
+            console.log(ops)
+            store.batch(ops, pc)
+        })
+        .catch(e => { console.log(e) })
+}
